fix: resolve circular import and add missing logical_inference parameter
- Extract shared types and logging functions to types_and_logging.py
- Update coordinator_batch to accept reasoning_engine parameter
- Pass reasoning_engine from coordinator wrapper to batch function
- This fixes the circular import between coordinator.py and coordinator_batch.py
- Ensures logical_inference parameter is properly handled
